+++ CHANGELOG.md
@@ -13,6 +13,7 @@ increment the patch number. The rational for versioning, and exact wording is ta
 - VEP now also produces a tab delimited file
 - CNVkit rules output genemetrics and gene break file
 - Added reference genome to be able to calculate AT/CG dropouts by Picard
+- coverage plot plugin part of issue #75
 
 ### Changed
 - Increased time for indel realigner and base recalib rules